Optimize neighbor array units
It is not possible for the type mask to be greater than 7 or the granularity to be greater than 64 on initialization. We can prove that easily and optimize accordingly.

+++ Source/Lib/Encoder/Codec/EbNeighborArrays.c
@@ -29,13 +29,13 @@ static void neighbor_array_unit_dctor32(EbPtr p) {
  *************************************************/
 EbErrorType svt_aom_neighbor_array_unit_ctor32(NeighborArrayUnit32 *na_unit_ptr, uint32_t max_picture_width,
                                                uint32_t max_picture_height, uint32_t unit_size,
-                                               uint32_t granularity_normal, uint32_t granularity_top_left,
-                                               uint32_t type_mask) {
+                                               uint8_t granularity_normal, uint8_t granularity_top_left,
+                                               uint8_t type_mask) {
     na_unit_ptr->dctor                     = neighbor_array_unit_dctor32;
     na_unit_ptr->unit_size                 = (uint8_t)(unit_size);
-    na_unit_ptr->granularity_normal        = (uint8_t)(granularity_normal);
+    na_unit_ptr->granularity_normal        = granularity_normal;
     na_unit_ptr->granularity_normal_log2   = (uint8_t)(svt_log2f(na_unit_ptr->granularity_normal));
-    na_unit_ptr->granularity_top_left      = (uint8_t)(granularity_top_left);
+    na_unit_ptr->granularity_top_left      = granularity_top_left;
     na_unit_ptr->granularity_top_left_log2 = (uint8_t)(svt_log2f(na_unit_ptr->granularity_top_left));
     na_unit_ptr->left_array_size           = (uint16_t)((type_mask & NEIGHBOR_ARRAY_UNIT_LEFT_MASK)
                                                             ? max_picture_height >> na_unit_ptr->granularity_normal_log2
@@ -71,13 +71,13 @@ static void neighbor_array_unit_dctor(EbPtr p) {
 
 EbErrorType svt_aom_neighbor_array_unit_ctor(NeighborArrayUnit *na_unit_ptr, uint32_t max_picture_width,
                                              uint32_t max_picture_height, uint32_t unit_size,
-                                             uint32_t granularity_normal, uint32_t granularity_top_left,
-                                             uint32_t type_mask) {
+                                             uint8_t granularity_normal, uint8_t granularity_top_left,
+                                             uint8_t type_mask) {
     na_unit_ptr->dctor                     = neighbor_array_unit_dctor;
     na_unit_ptr->unit_size                 = (uint8_t)(unit_size);
-    na_unit_ptr->granularity_normal        = (uint8_t)(granularity_normal);
+    na_unit_ptr->granularity_normal        = granularity_normal;
     na_unit_ptr->granularity_normal_log2   = (uint8_t)(svt_log2f(na_unit_ptr->granularity_normal));
-    na_unit_ptr->granularity_top_left      = (uint8_t)(granularity_top_left);
+    na_unit_ptr->granularity_top_left      = granularity_top_left;
     na_unit_ptr->granularity_top_left_log2 = (uint8_t)(svt_log2f(na_unit_ptr->granularity_top_left));
 
     na_unit_ptr->max_pic_h = max_picture_height;
@@ -284,7 +284,7 @@ void svt_aom_update_recon_neighbor_array16bit(NeighborArrayUnit *na_unit_ptr, ui
 void svt_aom_neighbor_array_unit_sample_write(NeighborArrayUnit *na_unit_ptr, uint8_t *src_ptr, uint32_t stride,
                                               uint32_t src_origin_x, uint32_t src_origin_y, uint32_t pic_origin_x,
                                               uint32_t pic_origin_y, uint32_t block_width, uint32_t block_height,
-                                              uint32_t neighbor_array_type_mask) {
+                                              uint8_t neighbor_array_type_mask) {
     uint32_t idx;
     uint8_t *dst_ptr;
     uint8_t *read_ptr;
@@ -426,7 +426,7 @@ void svt_aom_neighbor_array_unit_sample_write(NeighborArrayUnit *na_unit_ptr, ui
 void svt_aom_neighbor_array_unit16bit_sample_write(NeighborArrayUnit *na_unit_ptr, uint16_t *src_ptr, uint32_t stride,
                                                    uint32_t src_origin_x, uint32_t src_origin_y, uint32_t pic_origin_x,
                                                    uint32_t pic_origin_y, uint32_t block_width, uint32_t block_height,
-                                                   uint32_t neighbor_array_type_mask) {
+                                                   uint8_t neighbor_array_type_mask) {
     uint32_t  idx;
     uint16_t *dst_ptr;
     uint16_t *read_ptr;
@@ -574,7 +574,7 @@ void svt_aom_neighbor_array_unit16bit_sample_write(NeighborArrayUnit *na_unit_pt
  *************************************************/
 void svt_aom_neighbor_array_unit_mode_write32(NeighborArrayUnit32 *na_unit_ptr, uint32_t value, uint32_t org_x,
                                               uint32_t org_y, uint32_t block_width, uint32_t block_height,
-                                              uint32_t neighbor_array_type_mask) {
+                                              uint8_t neighbor_array_type_mask) {
     uint32_t  idx;
     uint32_t *dst_ptr;
 
@@ -683,7 +683,7 @@ void svt_aom_neighbor_array_unit_mode_write32(NeighborArrayUnit32 *na_unit_ptr, 
 
 void svt_aom_neighbor_array_unit_mode_write(NeighborArrayUnit *na_unit_ptr, uint8_t *value, uint32_t org_x,
                                             uint32_t org_y, uint32_t block_width, uint32_t block_height,
-                                            uint32_t neighbor_array_type_mask) {
+                                            uint8_t neighbor_array_type_mask) {
     uint32_t idx, j;
     uint8_t *dst_ptr;
 
@@ -794,7 +794,7 @@ void svt_aom_neighbor_array_unit_mode_write(NeighborArrayUnit *na_unit_ptr, uint
 }
 
 void svt_aom_copy_neigh_arr(NeighborArrayUnit *na_src, NeighborArrayUnit *na_dst, uint32_t org_x, uint32_t org_y,
-                            uint32_t bw, uint32_t bh, uint32_t neighbor_array_type_mask) {
+                            uint32_t bw, uint32_t bh, uint8_t neighbor_array_type_mask) {
     uint32_t idx;
     uint8_t *dst_ptr, *src_ptr;
 
@@ -859,7 +859,7 @@ void svt_aom_copy_neigh_arr(NeighborArrayUnit *na_src, NeighborArrayUnit *na_dst
 }
 
 void svt_aom_copy_neigh_arr_32(NeighborArrayUnit32 *na_src, NeighborArrayUnit32 *na_dst, uint32_t org_x, uint32_t org_y,
-                               uint32_t bw, uint32_t bh, uint32_t neighbor_array_type_mask) {
+                               uint32_t bw, uint32_t bh, uint8_t neighbor_array_type_mask) {
     uint32_t  idx;
     uint32_t *dst_ptr, *src_ptr;
